knowledge_service/requirements.txt
grpcio
grpcio-tools
protobuf
faiss-cpu
openai
numpy
python-dotenv
tqdm